+++ src/di/container.ts
@@ -14,16 +14,16 @@ import LocalUseCase from '@domain/local/use-cases/LocationUseCase'
 import { LocalCreateController } from '@presentation/http/controllers/local/CreateLocalController'
 import { ListLocalController } from '@presentation/http/controllers/local/ListLocalController'
 import { LocalRoutes } from '@presentation/http/routes/LocalRoutes'
-import PaymentMethodsRepository from '@domain/paymentMethods/repositories/PaymentMethodsRepository'
-import PaymentMethodsUseCase from '@domain/paymentMethods/use-cases/PaymentMethodsUseCase'
-import { CreatePaymentMethodsController } from '@presentation/http/controllers/paymentMethods/CreatePaymentMethodsController'
+import PaymentMethodsRepository from '@domain/payment-methods/repositories/PaymentMethodsRepository'
+import PaymentMethodsUseCase from '@domain/payment-methods/use-cases/PaymentMethodsUseCase'
+import { CreatePaymentMethodsController } from '@presentation/http/controllers/payment-methods/CreatePaymentMethodsController'
 import { PaymentMethodsRoutes } from '@presentation/http/routes/PaymentMethodsRoutes'
-import ServiceTypeRepository from '@domain/serviceType/repositories/ServiceTypeRepository'
-import ServiceTypeUseCase from '@domain/serviceType/use-cases/ServiceTypeUseCase'
-import { CreateServiceTypeController } from '@presentation/http/controllers/serviceType/CreateServiceTypeController'
+import ServiceTypeRepository from '@domain/service-type/repositories/ServiceTypeRepository'
+import ServiceTypeUseCase from '@domain/service-type/use-cases/ServiceTypeUseCase'
+import { CreateServiceTypeController } from '@presentation/http/controllers/service-type/CreateServiceTypeController'
 import { ServiceTypeRoutes } from '@presentation/http/routes/ServiceTypeRoutes'
-import { ListPaymentMethodsController } from '@presentation/http/controllers/paymentMethods/ListPaymentMethodsController'
-import { ListServiceTypeController } from '@presentation/http/controllers/serviceType/ListServiceTypeController'
+import { ListPaymentMethodsController } from '@presentation/http/controllers/payment-methods/ListPaymentMethodsController'
+import { ListServiceTypeController } from '@presentation/http/controllers/service-type/ListServiceTypeController'
 import { ListServicesController } from '@presentation/http/controllers/publications/services/ListServicesController'
 import PublicationTypeRepository from '@domain/publication-type/repositories/PublicationTypeRepository'
 import PublicationTypeUseCase from '@domain/publication-type/use-cases/PublicationTypeUseCase'
