+++ src/domain/service-type/interfaces/IServiceType.ts
