+++ src/domain/service-type/interfaces/IServiceTypeRepository.ts
