+++ src/domain/service-type/interfaces/IServiceTypeUseCase.ts
