+++ src/domain/service-type/repositories/ServiceTypeRepository.ts
